찐으로 좋아요 해결함

+++ src/api/feed.js
@@ -34,7 +34,7 @@ export async function addNewsfeedLike({ userId, feedId }) {
 	if (error) throw error;
 }
 
-export async function removNewsfeedLike({ feedId }) {
-	const { error } = await supabase.from("like").delete().eq("feedId", feedId);
+export async function removNewsfeedLike({ userId, feedId }) {
+	const { error } = await supabase.from("like").delete().eq("feedId", feedId).eq("userId", userId);
 	if (error) throw error;
 }
